Create backup_multi.bat
"Old school" script to backup multiple directories, copy to the backup server, and email when complete. This was used at a previous job to get around Backup Exec requirements for purchasing a client license for each server that needed backing up. Since BE could backup files on its local drive "for free", we cheated by compressing the files on the target server and copying to Backup Exec. (Moral: working for el cheapo employers makes you resourceful!)